fix(安全): 修复多个安全漏洞和内存泄漏问题

Created-by: spharx01
Commit-by: zhixianzhou
Merged-by: spharx01
Description: fix(安全): 修复多个安全漏洞和内存泄漏问题

See merge request: openairymax/agentos!109

+++ agentos/commons/utils/error/handler.c
@@ -670,7 +670,7 @@ const agentos_error_context_entry_t* agentos_error_chain_iter_next(
         return NULL;
     }
     
-    if (iter->current_index >= iter->chain->depth) {
+    if ((size_t)iter->current_index >= (size_t)iter->chain->depth) {
         return NULL;
     }
